update bikin folder models,controllers,route,database/migrations,app.js
