+++ package.json
@@ -16,6 +16,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "dompurify": "^3.2.1",
+    "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "lucide-react": "^0.456.0",
     "react": "^18.3.1",
